Allow dependencies with explicit extensions to skip metadata fetching
Allow dependencies declared with an explicit extension to skip metadata fetching (POM/IvyDescriptor).
This was implemented by updating the Default[Ivy,Maven]ArtifactRepository classes to always add a
modified version of the DefaultArtifactMetadataSource which returns metadata if the artifact's
definition contains an extension, thus skipping later metadata sources
which would otherwise fetch metadata from a pom/ivy descriptor.

This was not caught in the past, as the test we used to verify this
functionality was improperly written. It never published the test
artifact without a POM and continued to verify that the POM was being
fetched.